Split out tests so that env vars are not required for unit tests.
