+++ src/main/java/com/google/jenkins/plugins/k8sengine/client/ClientUtil.java
@@ -42,6 +42,7 @@ public class ClientUtil {
     Preconditions.checkNotNull(itemGroup);
     Preconditions.checkNotNull(domainRequirements);
     Preconditions.checkArgument(!Strings.isNullOrEmpty(credentialsId));
+    Preconditions.checkNotNull(transport);
 
     ClientFactory clientFactory;
     try {
