+++ src/test/resources/com/google/jenkins/plugins/k8sengine/client/test-key.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "test-project",
+  "private_key_id": "test-private-key-id",
+  "private_key": "test-private-key",
+  "client_email": "test-account@test-project.iam.gserviceaccount.com",
+  "client_id": "test-client-id",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/test-account%40test-project.iam.gserviceaccount.com"
+}
